pyproject.toml
[tool.poetry]
name = "scratchy-pad"
version = "0.1.0"
description = ""
authors = ["none <none>"]
readme = "README.md"
packages = [
    {include = "scratchy_pad_app"},
    {include = "scratchy_pad_client"},
    {include = "scripts"}
]

[tool.poetry.dependencies]
python = "^3.13"

scratchy_pad_domain = { path = "scratchy_pad_domain", develop = true }
scratchy_pad_adapters = { path = "scratchy_pad_adapters", develop = true }
grpcio = "^1.67.1"
grpcio-tools = "^1.67.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
serve = "scratchy_pad_app.main:run"
generate-grpc = "scripts.generate_grpc:generate"